+++ Encoder.h
@@ -28,6 +28,6 @@ void encoderInit();
 void getEncoderDistances(uint64_t * encoderDistances);
 
 // Gets the average number of ticks for the pair of encoders (used for PID of speed)
-void getAverageEncoderTicks(int64_t * avEncoderTicks);
+void getAverageEncoderTicks(int32_t * avEncoderTicks);
 
 #endif//_ENCODER_H_
